# Base16 optimizations