--- a/src/main/scala/scorex/crypto/authds/avltree/batch/ToStringHelper.scala
+++ b/src/main/scala/scorex/crypto/authds/avltree/batch/ToStringHelper.scala
@@ -1,8 +1,8 @@
 package scorex.crypto.authds.avltree.batch
 
-import scorex.utils.ScryptoLogging
+import scorex.utils.{ScorexEncoding, ScryptoLogging}
 
-trait ToStringHelper extends ScryptoLogging {
+trait ToStringHelper extends ScryptoLogging with ScorexEncoding {
   //Needed for debug (toString) only
   protected def arrayToString(a: Array[Byte]): String = encoder.encode(a).take(8)
 }