--- a/src/main/scala/scorex/crypto/encode/Base16.scala
+++ b/src/main/scala/scorex/crypto/encode/Base16.scala
@@ -1,19 +1,15 @@
 package scorex.crypto.encode
 
+import org.bouncycastle.util.encoders.Hex
+
 import scala.util.Try
 
 object Base16 extends BytesEncoder {
 
   override val Alphabet: String = "0123456789abcdefABCDEF"
 
-  def encode(input: Array[Byte]): String = bytes2hex(input, None)
+  def encode(input: Array[Byte]): String = Hex.toHexString(input)
 
-  def decode(input: String): Try[Array[Byte]] = Try(hex2bytes(input.toLowerCase()))
+  def decode(input: String): Try[Array[Byte]] = Try(Hex.decode(input))
 
-  private def bytes2hex(bytes: Array[Byte], sep: Option[String]): String =
-    bytes.map("%02x".format(_)).mkString(sep.getOrElse(""))
-
-  private def hex2bytes(hex: String): Array[Byte] = {
-    hex.sliding(2, 2).toArray.map(Integer.parseInt(_, 16).toByte)
-  }
 }