--- a/src/main/scala/scorex/utils/ScryptoLogging.scala
+++ b/src/main/scala/scorex/utils/ScryptoLogging.scala
@@ -1,10 +1,12 @@
 package scorex.utils
 
 import com.typesafe.scalalogging.StrictLogging
-import scorex.crypto.encode.{Base16, BytesEncoder}
 
+/**
+  * Trait with logger
+  * TODO extract to ScorexUtils
+  */
 trait ScryptoLogging extends StrictLogging {
   @inline protected def log = logger
 
-  implicit val encoder: BytesEncoder = Base16
 }