Before:
```
jmh:run -i 20 -wi 20 -f1 -t1 .*Base16*
Benchmark                                                 Mode  Cnt          Score          Error   Units
Base16Benchmark.decode                                   thrpt   20          4.021 ±        0.063   ops/s
Base16Benchmark.decode:·gc.alloc.rate.norm               thrpt   20  286390967.840 ±        8.344    B/op
Base16Benchmark.encode                                   thrpt   20          1.789 ±        0.018   ops/s
Base16Benchmark.encode:·gc.alloc.rate.norm               thrpt   20  672834978.400 ±        6.799    B/op
```
After
```
jmh:run -i 20 -wi 10 -f1 -t1 .*Base16*
Benchmark                                                 Mode  Cnt         Score        Error   Units
Base16Benchmark.decode                                   thrpt   20       152.509 ±     35.224   ops/s
Base16Benchmark.decode:·gc.alloc.rate.norm               thrpt   20   2768611.237 ±      0.999    B/op
Base16Benchmark.encode                                   thrpt   20       117.514 ±      2.537   ops/s
Base16Benchmark.encode:·gc.alloc.rate.norm               thrpt   20  10204515.674 ±      0.095    B/op
```